refactor: decompose FileAbilities into AgentFileAbilities + FlowFileAbilities
Split the monolithic FileAbilities (1,246 lines) and Files REST controller
(749 lines) into domain-specific classes aligned with the three-layer
memory architecture (site → agent identity → user).

New abilities layer:
- AgentFileAbilities: list/get/write/delete/upload for agent memory files
  with identity + user layer resolution
- FlowFileAbilities: list/get/delete/upload/cleanup for flow-scoped files

New REST controllers:
- AgentFiles: /files/agent/* and /files/agent/daily/* routes
- FlowFiles: /files (flow-scoped upload, list, delete)

Shared constants extracted to FileConstants (PROTECTED_FILES, USER_LAYER_FILES).

Old ability slugs replaced with domain-specific ones:
- datamachine/list-files → list-agent-files + list-flow-files
- datamachine/get-file → get-agent-file + get-flow-file
- datamachine/delete-file → delete-agent-file + delete-flow-file
- datamachine/upload-file → upload-agent-file + upload-flow-file
- datamachine/cleanup-files → cleanup-flow-files

Consumers updated: DeleteFile chat tool, Engine.php get_file_context,
bootstrap, all related tests.
