--- inc/Api/Files.php
@@ -1,749 +1,0 @@
-<?php
-/**
- * REST API Files Endpoint
- *
- * Unified file API supporting both flow-level files and pipeline-level context.
- * Delegates to FileAbilities for core logic.
- *
- * @package DataMachine\Api
- */
-
-namespace DataMachine\Api;
-
-use DataMachine\Abilities\PermissionHelper;
-use DataMachine\Abilities\DailyMemoryAbilities;
-use DataMachine\Abilities\FileAbilities;
-use WP_Error;
-use WP_REST_Request;
-use WP_REST_Response;
-use WP_REST_Server;
-
-if ( ! defined( 'WPINC' ) ) {
-	die;
-}
-
-class Files {
-
-	private static ?FileAbilities $abilities = null;
-
-	private static function getAbilities(): FileAbilities {
-		if ( null === self::$abilities ) {
-			self::$abilities = new FileAbilities();
-		}
-		return self::$abilities;
-	}
-
-	/**
-	 * Register REST API routes.
-	 */
-	public static function register() {
-		add_action( 'rest_api_init', array( self::class, 'register_routes' ) );
-	}
-
-	/**
-	 * Register /datamachine/v1/files endpoints.
-	 */
-	public static function register_routes() {
-		// POST /files - Upload file (flow or pipeline context)
-		register_rest_route(
-			'datamachine/v1',
-			'/files',
-			array(
-				'methods'             => WP_REST_Server::CREATABLE,
-				'callback'            => array( self::class, 'handle_upload' ),
-				'permission_callback' => array( self::class, 'check_permission' ),
-				'args'                => array(
-					'flow_step_id' => array(
-						'required'          => false,
-						'type'              => 'string',
-						'description'       => __( 'Flow step ID for flow-level files', 'data-machine' ),
-						'sanitize_callback' => function ( $param ) {
-							return sanitize_text_field( $param );
-						},
-					),
-				),
-			)
-		);
-
-		// GET /files - List files
-		register_rest_route(
-			'datamachine/v1',
-			'/files',
-			array(
-				'methods'             => WP_REST_Server::READABLE,
-				'callback'            => array( self::class, 'list_files' ),
-				'permission_callback' => array( self::class, 'check_permission' ),
-				'args'                => array(
-					'flow_step_id' => array(
-						'required'          => false,
-						'type'              => 'string',
-						'sanitize_callback' => function ( $param ) {
-							return sanitize_text_field( $param );
-						},
-					),
-				),
-			)
-		);
-
-		// DELETE /files/{filename} - Delete file
-		register_rest_route(
-			'datamachine/v1',
-			'/files/(?P<filename>[^/]+)',
-			array(
-				'methods'             => WP_REST_Server::DELETABLE,
-				'callback'            => array( self::class, 'delete_file' ),
-				'permission_callback' => array( self::class, 'check_permission' ),
-				'args'                => array(
-					'filename'     => array(
-						'required'          => true,
-						'type'              => 'string',
-						'sanitize_callback' => function ( $param ) {
-							return sanitize_file_name( $param );
-						},
-					),
-					'flow_step_id' => array(
-						'required'          => false,
-						'type'              => 'string',
-						'sanitize_callback' => function ( $param ) {
-							return sanitize_text_field( $param );
-						},
-					),
-				),
-			)
-		);
-
-		// GET /files/agent - List agent files
-		register_rest_route(
-			'datamachine/v1',
-			'/files/agent',
-			array(
-				'methods'             => WP_REST_Server::READABLE,
-				'callback'            => array( self::class, 'list_agent_files' ),
-				'permission_callback' => array( self::class, 'check_permission' ),
-				'args'                => array(
-					'user_id' => array(
-						'required'          => false,
-						'type'              => 'integer',
-						'sanitize_callback' => 'absint',
-					),
-				),
-			)
-		);
-
-		// GET /files/agent/{filename} - Get agent file content
-		register_rest_route(
-			'datamachine/v1',
-			'/files/agent/(?P<filename>[^/]+)',
-			array(
-				'methods'             => WP_REST_Server::READABLE,
-				'callback'            => array( self::class, 'get_agent_file' ),
-				'permission_callback' => array( self::class, 'check_permission' ),
-				'args'                => array(
-					'filename' => array(
-						'required'          => true,
-						'type'              => 'string',
-						'sanitize_callback' => function ( $param ) {
-							return sanitize_file_name( $param );
-						},
-					),
-					'user_id'  => array(
-						'required'          => false,
-						'type'              => 'integer',
-						'sanitize_callback' => 'absint',
-					),
-				),
-			)
-		);
-
-		// PUT /files/agent/{filename} - Write/update agent file content
-		register_rest_route(
-			'datamachine/v1',
-			'/files/agent/(?P<filename>[^/]+)',
-			array(
-				'methods'             => WP_REST_Server::EDITABLE,
-				'callback'            => array( self::class, 'put_agent_file' ),
-				'permission_callback' => array( self::class, 'check_permission' ),
-				'args'                => array(
-					'filename' => array(
-						'required'          => true,
-						'type'              => 'string',
-						'sanitize_callback' => function ( $param ) {
-							return sanitize_file_name( $param );
-						},
-					),
-					'user_id'  => array(
-						'required'          => false,
-						'type'              => 'integer',
-						'sanitize_callback' => 'absint',
-					),
-				),
-			)
-		);
-
-		// DELETE /files/agent/{filename} - Delete agent file
-		register_rest_route(
-			'datamachine/v1',
-			'/files/agent/(?P<filename>[^/]+)',
-			array(
-				'methods'             => WP_REST_Server::DELETABLE,
-				'callback'            => array( self::class, 'delete_agent_file' ),
-				'permission_callback' => array( self::class, 'check_permission' ),
-				'args'                => array(
-					'filename' => array(
-						'required'          => true,
-						'type'              => 'string',
-						'sanitize_callback' => function ( $param ) {
-							return sanitize_file_name( $param );
-						},
-					),
-					'user_id'  => array(
-						'required'          => false,
-						'type'              => 'integer',
-						'sanitize_callback' => 'absint',
-					),
-				),
-			)
-		);
-
-		// Daily memory file routes (YYYY/MM/DD convention).
-		$daily_date_args = array(
-			'user_id' => array(
-				'required'          => false,
-				'type'              => 'integer',
-				'sanitize_callback' => 'absint',
-			),
-			'year'    => array(
-				'required'          => true,
-				'type'              => 'string',
-				'validate_callback' => function ( $value ) {
-					return (bool) preg_match( '/^\d{4}$/', $value );
-				},
-			),
-			'month'   => array(
-				'required'          => true,
-				'type'              => 'string',
-				'validate_callback' => function ( $value ) {
-					return (bool) preg_match( '/^\d{2}$/', $value ) && (int) $value >= 1 && (int) $value <= 12;
-				},
-			),
-			'day'     => array(
-				'required'          => true,
-				'type'              => 'string',
-				'validate_callback' => function ( $value ) {
-					return (bool) preg_match( '/^\d{2}$/', $value ) && (int) $value >= 1 && (int) $value <= 31;
-				},
-			),
-		);
-
-		// GET /files/agent/daily - List daily memory files
-		register_rest_route(
-			'datamachine/v1',
-			'/files/agent/daily',
-			array(
-				'methods'             => WP_REST_Server::READABLE,
-				'callback'            => array( self::class, 'list_daily_files' ),
-				'permission_callback' => array( self::class, 'check_permission' ),
-				'args'                => array(
-					'user_id' => array(
-						'required'          => false,
-						'type'              => 'integer',
-						'sanitize_callback' => 'absint',
-					),
-				),
-			)
-		);
-
-		// GET /files/agent/daily/{year}/{month}/{day} - Read daily file
-		register_rest_route(
-			'datamachine/v1',
-			'/files/agent/daily/(?P<year>\d{4})/(?P<month>\d{2})/(?P<day>\d{2})',
-			array(
-				'methods'             => WP_REST_Server::READABLE,
-				'callback'            => array( self::class, 'get_daily_file' ),
-				'permission_callback' => array( self::class, 'check_permission' ),
-				'args'                => $daily_date_args,
-			)
-		);
-
-		// PUT /files/agent/daily/{year}/{month}/{day} - Write daily file
-		register_rest_route(
-			'datamachine/v1',
-			'/files/agent/daily/(?P<year>\d{4})/(?P<month>\d{2})/(?P<day>\d{2})',
-			array(
-				'methods'             => WP_REST_Server::EDITABLE,
-				'callback'            => array( self::class, 'put_daily_file' ),
-				'permission_callback' => array( self::class, 'check_permission' ),
-				'args'                => $daily_date_args,
-			)
-		);
-
-		// DELETE /files/agent/daily/{year}/{month}/{day} - Delete daily file
-		register_rest_route(
-			'datamachine/v1',
-			'/files/agent/daily/(?P<year>\d{4})/(?P<month>\d{2})/(?P<day>\d{2})',
-			array(
-				'methods'             => WP_REST_Server::DELETABLE,
-				'callback'            => array( self::class, 'delete_daily_file' ),
-				'permission_callback' => array( self::class, 'check_permission' ),
-				'args'                => $daily_date_args,
-			)
-		);
-	}
-
-	/**
-	 * Check user permission
-	 */
-	public static function check_permission( $request ) {
-		if ( ! is_user_logged_in() ) {
-			return new WP_Error(
-				'rest_forbidden',
-				__( 'You must be logged in to manage files.', 'data-machine' ),
-				array( 'status' => 401 )
-			);
-		}
-
-		$requested_user_id = self::resolve_scoped_user_id( $request );
-		$current_user_id   = get_current_user_id();
-
-		if ( $requested_user_id === $current_user_id ) {
-			return true;
-		}
-
-		if ( PermissionHelper::can( 'manage_agents' ) ) {
-			return true;
-		}
-
-		return new WP_Error(
-			'rest_forbidden',
-			__( 'You do not have permission to access another user\'s files.', 'data-machine' ),
-			array( 'status' => 403 )
-		);
-	}
-
-	/**
-	 * Resolve the effective REST-scoped user ID.
-	 *
-	 * If user_id is omitted, REST requests default to the current user instead
-	 * of the shared/default agent context.
-	 *
-	 * @param WP_REST_Request $request REST request.
-	 * @return int
-	 */
-	private static function resolve_scoped_user_id( WP_REST_Request $request ): int {
-		$user_id = $request->get_param( 'user_id' );
-
-		if ( null !== $user_id && '' !== $user_id ) {
-			return (int) $user_id;
-		}
-
-		return get_current_user_id();
-	}
-
-	/**
-	 * List files for flow or pipeline context
-	 */
-	public static function list_files( WP_REST_Request $request ) {
-		$flow_step_id = $request->get_param( 'flow_step_id' );
-
-		$input = array();
-
-		if ( $flow_step_id ) {
-			$input['flow_step_id'] = sanitize_text_field( $flow_step_id );
-		}
-
-		$result = self::getAbilities()->executeListFiles( $input );
-
-		if ( ! $result['success'] ) {
-			$status = 400;
-			if ( false !== strpos( $result['error'] ?? '', 'not found' ) ) {
-				$status = 404;
-			}
-
-			return new WP_Error(
-				'list_files_error',
-				$result['error'],
-				array( 'status' => $status )
-			);
-		}
-
-		return rest_ensure_response(
-			array(
-				'success' => true,
-				'data'    => $result['files'],
-			)
-		);
-	}
-
-	/**
-	 * Handle file upload for flow files or pipeline context.
-	 * Delegates to FileAbilities::executeUploadFile().
-	 */
-	public static function handle_upload( WP_REST_Request $request ) {
-		$flow_step_id = $request->get_param( 'flow_step_id' );
-
-		$files = $request->get_file_params();
-		if ( empty( $files['file'] ) || ! is_array( $files['file'] ) ) {
-			return new WP_Error(
-				'missing_file',
-				__( 'File upload is required.', 'data-machine' ),
-				array( 'status' => 400 )
-			);
-		}
-
-		$uploaded = $files['file'];
-		$input    = array(
-			'file_data' => array(
-				'name'     => sanitize_file_name( $uploaded['name'] ?? '' ),
-				'type'     => sanitize_mime_type( $uploaded['type'] ?? '' ),
-				'tmp_name' => $uploaded['tmp_name'] ?? '',
-				'error'    => intval( $uploaded['error'] ?? UPLOAD_ERR_NO_FILE ),
-				'size'     => intval( $uploaded['size'] ?? 0 ),
-			),
-		);
-
-		if ( $flow_step_id ) {
-			$input['flow_step_id'] = sanitize_text_field( $flow_step_id );
-		}
-
-		$result = self::getAbilities()->executeUploadFile( $input );
-
-		if ( ! $result['success'] ) {
-			$status = 400;
-			if ( false !== strpos( $result['error'] ?? '', 'not found' ) ) {
-				$status = 404;
-			} elseif ( false !== strpos( $result['error'] ?? '', 'Failed to store' ) ) {
-				$status = 500;
-			}
-
-			return new WP_Error(
-				'upload_file_error',
-				$result['error'],
-				array( 'status' => $status )
-			);
-		}
-
-		return rest_ensure_response(
-			array(
-				'success' => true,
-				'data'    => array(
-					'files' => $result['files'],
-					'scope' => $result['scope'],
-				),
-			)
-		);
-	}
-
-	/**
-	 * Delete file (flow or pipeline context)
-	 */
-	public static function delete_file( WP_REST_Request $request ) {
-		$filename     = sanitize_file_name( wp_unslash( $request['filename'] ) );
-		$flow_step_id = $request->get_param( 'flow_step_id' );
-
-		$input = array(
-			'filename' => $filename,
-		);
-
-		if ( $flow_step_id ) {
-			$input['flow_step_id'] = sanitize_text_field( $flow_step_id );
-		}
-
-		$result = self::getAbilities()->executeDeleteFile( $input );
-
-		if ( ! $result['success'] ) {
-			$status = 400;
-			if ( false !== strpos( $result['error'] ?? '', 'not found' ) ) {
-				$status = 404;
-			}
-
-			return new WP_Error(
-				'delete_file_error',
-				$result['error'],
-				array( 'status' => $status )
-			);
-		}
-
-		return rest_ensure_response(
-			array(
-				'success' => true,
-				'data'    => array(
-					'scope' => $result['scope'],
-				),
-			)
-		);
-	}
-
-	/**
-	 * List agent files.
-	 */
-	public static function list_agent_files( WP_REST_Request $request ) {
-		$input = array(
-			'scope'   => 'agent',
-			'user_id' => self::resolve_scoped_user_id( $request ),
-		);
-
-		$result = self::getAbilities()->executeListFiles( $input );
-
-		if ( ! $result['success'] ) {
-			return new WP_Error( 'list_agent_files_error', $result['error'], array( 'status' => 500 ) );
-		}
-
-		return rest_ensure_response(
-			array(
-				'success' => true,
-				'data'    => $result['files'],
-			)
-		);
-	}
-
-	/**
-	 * Get agent file content.
-	 */
-	public static function get_agent_file( WP_REST_Request $request ) {
-		$filename = sanitize_file_name( wp_unslash( $request['filename'] ) );
-
-		$input = array(
-			'filename' => $filename,
-			'scope'    => 'agent',
-			'user_id'  => self::resolve_scoped_user_id( $request ),
-		);
-
-		$result = self::getAbilities()->executeGetFile( $input );
-
-		if ( ! $result['success'] ) {
-			$status = false !== strpos( $result['error'] ?? '', 'not found' ) ? 404 : 400;
-			return new WP_Error( 'get_agent_file_error', $result['error'], array( 'status' => $status ) );
-		}
-
-		return rest_ensure_response(
-			array(
-				'success' => true,
-				'data'    => $result['file'],
-			)
-		);
-	}
-
-	/**
-	 * Write/update agent file content (accepts raw body as content).
-	 * Delegates to FileAbilities::executeWriteAgentFile().
-	 */
-	public static function put_agent_file( WP_REST_Request $request ) {
-		$filename = sanitize_file_name( wp_unslash( $request['filename'] ) );
-		$content  = $request->get_body();
-
-		$input = array(
-			'filename' => $filename,
-			'content'  => $content,
-			'user_id'  => self::resolve_scoped_user_id( $request ),
-		);
-
-		$result = self::getAbilities()->executeWriteAgentFile( $input );
-
-		if ( ! $result['success'] ) {
-			$status = 400;
-			if ( false !== strpos( $result['error'] ?? '', 'Filesystem' ) || false !== strpos( $result['error'] ?? '', 'Failed' ) ) {
-				$status = 500;
-			}
-			return new WP_Error( 'put_agent_file_error', $result['error'], array( 'status' => $status ) );
-		}
-
-		return rest_ensure_response(
-			array(
-				'success'  => true,
-				'filename' => $result['filename'],
-			)
-		);
-	}
-
-	/**
-	 * Delete agent file.
-	 *
-	 * Defense-in-depth: checks protected files at REST layer before delegating to abilities.
-	 */
-	public static function delete_agent_file( WP_REST_Request $request ) {
-		$filename = sanitize_file_name( wp_unslash( $request['filename'] ) );
-
-		// Defense-in-depth: block deletion of protected files at the REST layer too.
-		if ( in_array( $filename, FileAbilities::PROTECTED_FILES, true ) ) {
-			return new WP_Error(
-				'delete_agent_file_error',
-				sprintf( 'Cannot delete protected file: %s', $filename ),
-				array( 'status' => 403 )
-			);
-		}
-
-		$input = array(
-			'filename' => $filename,
-			'scope'    => 'agent',
-			'user_id'  => self::resolve_scoped_user_id( $request ),
-		);
-
-		$result = self::getAbilities()->executeDeleteFile( $input );
-
-		if ( ! $result['success'] ) {
-			$status = false !== strpos( $result['error'] ?? '', 'not found' ) ? 404 : 400;
-			return new WP_Error( 'delete_agent_file_error', $result['error'], array( 'status' => $status ) );
-		}
-
-		return rest_ensure_response(
-			array(
-				'success' => true,
-				'data'    => array( 'scope' => $result['scope'] ),
-			)
-		);
-	}
-
-	// =========================================================================
-	// Daily Memory Handlers
-	// =========================================================================
-
-	/**
-	 * List all daily memory files grouped by month.
-	 * Delegates to DailyMemoryAbilities::listDaily().
-	 *
-	 * @since 0.32.0
-	 */
-	public static function list_daily_files( WP_REST_Request $request ) {
-		$input = array(
-			'user_id' => self::resolve_scoped_user_id( $request ),
-		);
-
-		$result = DailyMemoryAbilities::listDaily( $input );
-
-		return rest_ensure_response(
-			array(
-				'success' => true,
-				'data'    => $result['months'],
-			)
-		);
-	}
-
-	/**
-	 * Get a daily memory file's content.
-	 * Delegates to DailyMemoryAbilities::readDaily().
-	 *
-	 * @since 0.32.0
-	 */
-	public static function get_daily_file( WP_REST_Request $request ) {
-		$date = sprintf( '%s-%s-%s', $request['year'], $request['month'], $request['day'] );
-
-		$input = array(
-			'date'    => $date,
-			'user_id' => self::resolve_scoped_user_id( $request ),
-		);
-
-		$result = DailyMemoryAbilities::readDaily( $input );
-
-		if ( ! $result['success'] ) {
-			return new WP_Error( 'daily_file_not_found', $result['message'], array( 'status' => 404 ) );
-		}
-
-		return rest_ensure_response(
-			array(
-				'success' => true,
-				'data'    => array(
-					'date'    => $result['date'],
-					'content' => $result['content'],
-				),
-			)
-		);
-	}
-
-	/**
-	 * Write/update a daily memory file (accepts raw body as content).
-	 * Delegates to DailyMemoryAbilities::writeDaily().
-	 *
-	 * Respects the daily_memory_enabled setting via the abilities layer.
-	 *
-	 * @since 0.32.0
-	 */
-	public static function put_daily_file( WP_REST_Request $request ) {
-		$date    = sprintf( '%s-%s-%s', $request['year'], $request['month'], $request['day'] );
-		$content = $request->get_body();
-
-		$input = array(
-			'date'    => $date,
-			'content' => $content,
-			'mode'    => 'write',
-			'user_id' => self::resolve_scoped_user_id( $request ),
-		);
-
-		$result = DailyMemoryAbilities::writeDaily( $input );
-
-		if ( ! $result['success'] ) {
-			$status = false !== strpos( $result['message'] ?? '', 'disabled' ) ? 403 : 500;
-			return new WP_Error( 'daily_file_write_error', $result['message'], array( 'status' => $status ) );
-		}
-
-		return rest_ensure_response(
-			array(
-				'success' => true,
-				'date'    => $date,
-				'message' => $result['message'],
-			)
-		);
-	}
-
-	/**
-	 * Delete a daily memory file.
-	 * Delegates to DailyMemoryAbilities::deleteDaily().
-	 *
-	 * Respects the daily_memory_enabled setting via the abilities layer.
-	 *
-	 * @since 0.32.0
-	 */
-	public static function delete_daily_file( WP_REST_Request $request ) {
-		$date = sprintf( '%s-%s-%s', $request['year'], $request['month'], $request['day'] );
-
-		$input = array(
-			'date'    => $date,
-			'user_id' => self::resolve_scoped_user_id( $request ),
-		);
-
-		$result = DailyMemoryAbilities::deleteDaily( $input );
-
-		if ( ! $result['success'] ) {
-			$status = false !== strpos( $result['message'] ?? '', 'disabled' ) ? 403 : 404;
-			return new WP_Error( 'daily_file_delete_error', $result['message'], array( 'status' => $status ) );
-		}
-
-		return rest_ensure_response(
-			array(
-				'success' => true,
-				'message' => $result['message'],
-			)
-		);
-	}
-
-	/**
-	 * Get file context array from flow ID.
-	 *
-	 * Supports both database flows (numeric ID) and direct execution ('direct').
-	 *
-	 * @param int|string $flow_id Flow ID or 'direct' for ephemeral workflows
-	 * @return array Context array with pipeline_id and flow_id
-	 */
-	public static function get_file_context( int|string $flow_id ): array {
-		// Direct execution mode - no database lookup needed
-		if ( 'direct' === $flow_id ) {
-			return array(
-				'pipeline_id' => 'direct',
-				'flow_id'     => 'direct',
-			);
-		}
-
-		$db_flows  = new \DataMachine\Core\Database\Flows\Flows();
-		$flow_data = $db_flows->get_flow( (int) $flow_id );
-
-		if ( ! isset( $flow_data['pipeline_id'] ) || empty( $flow_data['pipeline_id'] ) ) {
-			throw new \InvalidArgumentException( 'Flow data missing required pipeline_id' );
-		}
-
-		$pipeline_id = $flow_data['pipeline_id'];
-
-		return array(
-			'pipeline_id' => $pipeline_id,
-			'flow_id'     => $flow_id,
-		);
-	}
-}
